+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * Add ACF's Textarea and Wysiwyg support.
 * Add WP CLI command to list and delete unused media.
 * Finally lower PHP requirements to 5.6.
+* Fix wp-cli command registering.
+* Improve speed of wp-cli.
+* Add wp-cli progress bar for indexation command.
 
 ## 2.0.3 - 04 Feb 2018
 * Go to WP.Org.